# services/gateway/docs/swagger.yaml
definitions:
...
  github_com_ummuys_reportify_services_gateway_internal_webdto.CreateUserRequest:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.CreateUserResponse:
    properties:
      user_id:
        type: string
    type: object
...
  github_com_ummuys_reportify_services_gateway_internal_webdto.DeleteUserResponse:
    properties:
      user_id:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse:
    properties:
      err:
        type: string
    type: object
...
  github_com_ummuys_reportify_services_gateway_internal_webdto.ListUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.User'
        type: array
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.LoginResponse:
    properties:
      access_token:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.RefreshTokenResponse:
    properties:
      access_token:
        type: string
    type: object
...
  github_com_ummuys_reportify_services_gateway_internal_webdto.UpdateUserRequest:
    properties:
      is_active:
        type: boolean
      password:
        type: string
      role:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.UpdateUserResponse:
    properties:
      is_active:
        type: boolean
      role:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  github_com_ummuys_reportify_services_gateway_internal_webdto.User:
    properties:
      is_active:
        type: boolean
      role:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and sets refresh token cookie. Returns access
        token.
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      summary: Login
      tags:
      - auth
  /auth/refresh:
    get:
      description: Uses refresh_token cookie to issue a new access token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.RefreshTokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      summary: Refresh access token
      tags:
      - auth
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ListUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - admin
  /users/{user_id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.DeleteUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ummuys_reportify_services_gateway_internal_webdto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - admin
swagger: "2.0"